chore(provider/fireworks): update Kimi examples to K2.6
