+++ .changeset/fresh-fireworks-cache.md
@@ -1,5 +1,6 @@
 ---
-'@ai-sdk/fireworks': patch
+"@ai-sdk/fireworks": patch
 ---
 
-Add the `promptCacheKey` provider option for Fireworks prompt cache affinity.
+Add the `promptCacheKey` provider option for Fireworks prompt cache affinity
+and Kimi K2.6 model autocomplete.
