+++ examples/ai-functions/src/generate-text/fireworks/kimi-k2-6-tool-call.ts
@@ -2,16 +2,15 @@ import {
   fireworks,
   type FireworksLanguageModelOptions,
 } from '@ai-sdk/fireworks';
-import { isStepCount, streamText } from 'ai';
-import { printFullStream } from '../../lib/print-full-stream';
+import { generateText, isStepCount } from 'ai';
 import { run } from '../../lib/run';
 import { weatherTool } from '../../tools/weather-tool';
 
 run(async () => {
   const sessionId = 'weather-conversation-123';
 
-  const result = streamText({
-    model: fireworks('accounts/fireworks/models/kimi-k2p5'),
+  const result = await generateText({
+    model: fireworks('accounts/fireworks/models/kimi-k2p6'),
     providerOptions: {
       fireworks: {
         promptCacheKey: sessionId,
@@ -24,6 +23,12 @@ run(async () => {
     prompt: 'What is the weather in San Francisco?',
   });
 
-  await printFullStream({ result });
-  console.log('Token usage:', await result.usage);
+  console.log(result.text);
+  console.log();
+  console.log('Token usage:', result.usage);
+  console.log(
+    'Cached input tokens:',
+    result.usage.inputTokenDetails.cacheReadTokens,
+  );
+  console.log('Finish reason:', result.finishReason);
 });
