+++ packages/fireworks/src/fireworks-chat-options.ts
@@ -21,6 +21,7 @@ export type FireworksChatModelId =
   | 'accounts/fireworks/models/kimi-k2-instruct'
   | 'accounts/fireworks/models/kimi-k2-thinking'
   | 'accounts/fireworks/models/kimi-k2p5'
+  | 'accounts/fireworks/models/kimi-k2p6'
   | 'accounts/fireworks/models/minimax-m2'
   | (string & {});
